services/api-server/pyproject.toml
[project]
name = "api-server"
version = "0.1.0"
description = "FastAPI REST + WebSocket API for Kutana AI"
requires-python = ">=3.12"
dependencies = [
    "kutana-core",
    "agent-gateway",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "redis>=5.0",
    "pydantic-settings>=2.0",
    "bcrypt>=4.0",
    "PyJWT>=2.8",
    "python-multipart>=0.0.9",
    "email-validator>=2.0",
    "stripe>=15.0.1",
    "boto3>=1.35",
    "sentry-sdk[fastapi]>=2.0",
    "structlog>=24.0",
    "prometheus-fastapi-instrumentator>=7.0",
    "aiohttp>=3.9",
    "aiosmtplib>=3.0",
    "anthropic>=0.84.0",
    "langfuse>=3.0",
    "livekit-api>=0.7,<1.0",
]

[tool.uv.sources]
kutana-core = { workspace = true }
agent-gateway = { workspace = true }

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/api_server"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
